+++ pkg/mimo/errors.go
@@ -28,4 +28,10 @@ var (
 
 	// ErrKeyNotFound is returned when key is not in multimap.
 	ErrKeyNotFound = errors.New("key not found in multimap")
+
+	// ErrDisparityType is returned when real and masked value have different types.
+	ErrDisparityType = errors.New("can't validate value because type has changed")
+
+	// ErrAnalyze is returned when a fatal error occurs during analyze.
+	ErrAnalyze = errors.New("fatal error")
 )